requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.0.0
